# [MTG-1166] Changing owner index to speed up owner based requests